frontend/requirements.txt
Flask
python-chess
requests